chore: Move `enable` to request. Create `_checkAndChangeNetwork`
