+++ source/scripts/Background/controllers/message-handler/index.ts
@@ -1,38 +1,7 @@
 import { browser, Runtime } from 'webextension-polyfill-ts';
 
-import store from 'state/store';
-
-import { popupPromise } from './popup-promise';
-import { methodRequest } from './requests';
+import { methodRequest, enable } from './requests';
 import { Message } from './types';
-
-const enable = async (host: string, data: any) => {
-  const { chain, chainId } = data;
-  const { dapp, wallet } = window.controller;
-
-  if (dapp.isConnected(host)) return { success: true };
-
-  // network check
-  const {
-    vault: { activeNetwork, networks },
-  } = store.getState();
-  const isSysCore = activeNetwork.url.includes('blockbook');
-  const activeChain = isSysCore ? 'syscoin' : 'ethereum';
-
-  const isSameChain = chain === activeChain;
-  const isSameChainId = activeNetwork.chainId === chainId;
-  if (!isSameChain || !isSameChainId) {
-    const network = networks[chain][chainId];
-    await wallet.setActiveNetwork(network);
-  }
-
-  return popupPromise({
-    host,
-    route: 'connect-wallet',
-    eventName: 'connect',
-    data: { chain, chainId },
-  });
-};
 
 /**
  * Handles:
